PSG refactoring
- break out the constraint routine into the AscentProblem
- implement Kepler5 in the Ascent API
- do some more AutoDiff work and removing hand done derivatives
- disable the optguard derivative checks being a DEBUG flag
