+++ MechJebLibTest/PSGTests/AscentTests/BuggyTests.cs
@@ -35,7 +35,7 @@ namespace MechJebLibTest.PSGTests.AscentTests
             double attR  = rbody + 250000;
             Ascent ascent = Ascent.Builder()
                 .Initial(r0, v0, u0, t0, mu, rbody)
-                .SetTarget(attR, attR, attR, 0.498274048151861, 0, 0, true, false)
+                .SetTarget(attR, attR, attR, 0.498274048151861, 0, 0, 0, true, false, false)
                 .AddStageUsingFinalMass(23796.5132378765, 3243.62653856172, 241.889217870454, 127.294984318196, 4, 4)
                 .AddCoast(438.273627540912, 0, 450, 3, 3)
                 .AddStageUsingFinalMass(438.273627540912, 200.596948133534, 220.000025355962, 5.82703026766676, 2, 2, true, false)
@@ -75,7 +75,7 @@ namespace MechJebLibTest.PSGTests.AscentTests
 
             Ascent ascent = Ascent.Builder()
                 .Initial(r0, v0, u0, t0, mu, rbody)
-                .SetTarget(6556000, 6557000, 6556000, 0.499303792410538, 0, 0, true, false)
+                .SetTarget(6556000, 6557000, 6556000, 0.499303792410538, 0, 0, 0, true, false, false)
                 .AddStageUsingFinalMass(52831.1138786043, 6373.37509359599, 287.450291861922, 187.198692023839, 3, 3, allowShutdown: false)
                 .AddStageUsingFinalMass(2517.01227921406, 738.314292468521, 270.157637827755, 139.419410735907, 1, 1, allowShutdown: false)
                 .AddStageUsingFinalMass(81.5125042572618, 74.8691843417473, 250.000044703484, 0.678632409425986, 0, 0, true, false)
@@ -127,7 +127,7 @@ namespace MechJebLibTest.PSGTests.AscentTests
 
             Ascent ascent = Ascent.Builder()
                 .Initial(r0, v0, u0, t0, mu, rbody)
-                .SetTarget(6571000, 6571000, 6571000, 0.558505360638185, 0, 0.349065850398866, true, false)
+                .SetTarget(6571000, 6571000, 6571000, 0.558505360638185, 0, 0, 0.349065850398866, true, false, false)
                 .AddStageUsingFinalMass(731995.326793174 - decmass, 328918.196876464 - decmass, 408.091742854086, 159.080051443926, 4, 4)
                 .AddStageUsingFinalMass(268744.821741949 - decmass, 168530.570801559 - decmass, 408.091702159863, 118.652885602609, 2, 2)
                 .AddStageUsingFinalMass(40763.8839289468 - decmass, 13796.4420050909 - decmass, 465.500076480742, 1118.13132581707, 1, 1)
@@ -156,7 +156,7 @@ namespace MechJebLibTest.PSGTests.AscentTests
 
             Ascent ascent = Ascent.Builder()
                 .Initial(r0, v0, u0, t0, mu, rbody)
-                .SetTarget(6621000, 6623000, 6621000, 0.558505360638185, 0, 0.349065850398866, false, false)
+                .SetTarget(6621000, 6623000, 6621000, 0.558505360638185, 0, 0, 0.349065850398866, false, false, false)
                 .AddStageUsingFinalMass(51814.6957082437, 12381.5182945184, 277.252333393375, 139.843831752395, 2, 2)
                 .AddStageUsingFinalMass(9232.92402212159, 887.216491554419, 252.019434034823, 529.378964006269, 1, 1)
                 .Build();
@@ -179,7 +179,7 @@ namespace MechJebLibTest.PSGTests.AscentTests
 
             Ascent ascent = Ascent.Builder()
                 .Initial(new V3(-2750001.81262551, -4870536.29021665, 3050610.57762817), new V3(355.169076284135, -200.527924207868, -0.00280072807318912), new V3(-0.431553114381181, -0.764704888551875, 0.478527367115021), 139051.267268641, 398600435436096, 6371000)
-                .SetTarget(6551000, 6551000, 6551000, 0.499303792410538, 0, 0, false, false)
+                .SetTarget(6551000, 6551000, 6551000, 0.499303792410538, 0, 0, 0, false, false, false)
                 //.SetTarget(8051000, 8051000, 8051000, 0.499303792410538, 0, 0, false, false)
                 .AddStageUsingFinalMass(176568.069166482, 33739.9978121645, 294.100105371403, 150.314019929573, 7, 7)
                 .AddStageUsingFinalMass(22640.1444759873, 7657.9646078915, 427.000116464637, 470.292183477623, 4, 4)
@@ -214,7 +214,7 @@ namespace MechJebLibTest.PSGTests.AscentTests
 
             Ascent ascent = Ascent.Builder()
                 .Initial(new V3(-3255967.18116409, -4547891.60302231, 3050610.57695918), new V3(331.637492561945, -237.428880310986, 7.25192369772431E-06), new V3(-0.511086312614353, -0.713789180806743, 0.478848457336426), 309907.485302214, 398600435436096, 6371000)
-                .SetTarget(PER, 6371000, PER, 0.499303792410538, 0, 0, false, false)
+                .SetTarget(PER, 6371000, PER, 0.499303792410538, 0, 0, 0, false, false, false)
                 .AddStageUsingFinalMass(176568.069166482, 33739.9978121645, 294.100108674295, 150.314019929573, 7, 7)
                 .AddStageUsingFinalMass(22640.1444759873, 7657.9646078915, 427.000108671115, 470.292183477623, 4, 4)
                 .AddCoast(529.255624515231, 200, 2000, 1, 0)
@@ -230,7 +230,7 @@ namespace MechJebLibTest.PSGTests.AscentTests
 
             Ascent ascent2 = Ascent.Builder()
                 .Initial(new V3(-3255967.18116409, -4547891.60302231, 3050610.57695918), new V3(331.637492561945, -237.428880310986, 7.25192369772431E-06), new V3(-0.511086312614353, -0.713789180806743, 0.478848457336426), 309907.485302214, 398600435436096, 6371000)
-                .SetTarget(PER, 6371000, PER, 0.499303792410538, 0, 0, false, false)
+                .SetTarget(PER, 6371000, PER, 0.499303792410538, 0, 0, 0, false, false, false)
                 .AddStageUsingFinalMass(176568.069166482, 33739.9978121645, 294.100108674295, 150.314019929573, 7, 7)
                 .AddStageUsingFinalMass(22640.1444759873, 7657.9646078915, 427.000108671115, 470.292183477623, 4, 4)
                 .AddCoast(529.255624515231, 200, 2000, 1, 0)
