+++ MechJebLibTest/PSGTests/AscentTests/TheStandardTests.cs
@@ -38,7 +38,7 @@ namespace MechJebLibTest.PSGTests.AscentTests
 
             Ascent ascent = Ascent.Builder()
                 .Initial(r0, v0, u0, t0, mu, rbody)
-                .SetTarget(PeR, ApR, PeR, incT, 0, 0, true, false)
+                .SetTarget(PeR, ApR, PeR, incT, 0, 0, 0, true, false, false)
                 .AddStageUsingFinalMass(49119.7842689869, 7114.2513992454, 288.000034332275, 170.308460385726, 3, 3)
                 .AddStageUsingFinalMass(2848.62586760223, 1363.71123994759, 270.15767003304, 116.391834883409, 1, 1)
                 .AddCoast(678.290157913434, 0, 450, 1, 1)
@@ -99,7 +99,7 @@ namespace MechJebLibTest.PSGTests.AscentTests
 
             Ascent ascent = Ascent.Builder()
                 .Initial(r0, v0, u0, t0, mu, rbody)
-                .SetTarget(PeR, ApR, PeR, incT, 0, 0, false, false)
+                .SetTarget(PeR, ApR, PeR, incT, 0, 0, 0, false, false, false)
                 .AddStageUsingFinalMass(49119.7842689869, 7114.2513992454, 288.000034332275, 170.308460385726, 3, 3)
                 .AddStageUsingFinalMass(2848.62586760223, 1363.71123994759, 270.15767003304, 116.391834883409, 1, 1)
                 .AddCoast(678.290157913434, 0, 450, 1, 1)
@@ -160,7 +160,7 @@ namespace MechJebLibTest.PSGTests.AscentTests
 
             Ascent ascent = Ascent.Builder()
                 .Initial(r0, v0, u0, t0, mu, rbody)
-                .SetTarget(PeR, ApR, PeR, incT, 0, 0, true, false)
+                .SetTarget(PeR, ApR, PeR, incT, 0, 0, 0, true, false, false)
                 .AddStageUsingFinalMass(49119.7842689869, 7114.2513992454, 288.000034332275, 170.308460385726, 3, 3)
                 .AddStageUsingFinalMass(2848.62586760223, 1363.71123994759, 270.15767003304, 116.391834883409, 1, 1)
                 .AddCoast(678.290157913434, 0, 450, 1, 1)
